Refactor of rest_svc.py
