+++ tests/test_iocs.py
@@ -2,7 +2,7 @@ import os
 import random
 
 from tests.thread_app_test import ThreadAppTest
-from threadcomponents.service.rest_svc import UID as UID_KEY
+from threadcomponents.constants import UID as UID_KEY
 from uuid import uuid4
