+++ threadcomponents/constants.py
@@ -1,0 +1,7 @@
+REST_IGNORED = dict(ignored=1)
+REST_SUCCESS = dict(success=1)
+
+DATETIME_OBJ = "datetime_obj"
+UID = "uid"
+URL = "url"
+TITLE = "title"
